+++ src/components/ratings/Ratings.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    ratings done
+  </div>
+</template>
+
+<script>
+export default {
+};
+</script>
+
+<style>
+</style>